statistics::product_analytics: Switch to monthly execution
We've verified that systemd_timer and the test notebook are
executed as expected, so we are good to go on updating the
interval to monthly cadence.

Executing on the the 9th of every month helps ensure that a
new MediaWiki History snapshot has been generated, since
those usually take no more than 8 days.

Bug: T291957
Change-Id: I6f5f87d22a4914b312898b8270e752bc60dc89f5