wikistats: drop unneeded wikistats_host parameter
The _host parameter is not even used anymore, was just
in commented out rewrite rule from the past
and used for the ServerName and resource name in httpd.

But with that being behind proxy in cloud VPS it is rather
meaningless.

So might as well simplify and drop it altogether since
currently there is a code issue with it not being found with
a lookup in Hiera.

Change-Id: I2d7a8a4bd85ad1207bea8e797c00a1346d2ccb45