profile::kafka::broker: add new Kafka PKI intermediate CA option
We created a new PKI Intermediate CA for Kafka, and this change
adds the necessary code on the brokers' profile to generate
their hostname-based TLS certificates (instead of re-using
or exposing the puppet ones). The idea is to have a flag to
force the creation of the new keytstores/truststores on selected
clusters and allow a smooth transition.

This change will also need to be followed by client-related
ones, it will not be sufficient of course, but it is a good
starting point to experiment in the Kafka test cluster.

Assumption: the trustore file is not mandatory since
we can rely on the default jvm trustore. If Kafka requires
it to run any TLS config, we'll need to follow up.

Bug: T291905
Change-Id: I6f56e31973816185c268db810547faa22af6496a