tlsproxy::localssl: Allow disabling http2
Why:
The module is still being used, in both hosts/services with public IPs
as well as internal hosts and services. While for end users, in most
cases http2 is beneficial, internal services are a different beast and
the decision to enable http2 should be a conscious one. Unfortunately
when 125f28e452aa was committed, all users of the module got http2 enabled
unconditionally. This made a lot of sense for the edge caches back then
and the users of this code were fewer. This mostly was harmless because
internal clients were still preferring http1.0 or http1.1, so this went
unnoticed.  With the switch to buster, libcurl defaults to http2 by
default and thus mediawiki switched to http2 as well for uses of
MultiHttpClient. Again, this in most cases is ok, until T275752 was
noticed, with heavy performance degradations for large file size
uploads. Having http2 in this case proved detrimental due to a number of
factors.

What:
* Allow to conditionally enable http2 in tlsproxy::localssl
* Maintain the current status quo, by forcing to true the new
  conditional in all users, with the exception of swift

Followup patches will disable http2 for others users. Service Owners are
free to enable it, but it should be a conscious decision after pros and
cons have been weighted.

It's important to note that the code isn't being used any more by the
edge caches as the traffic team has switch over to ats-tls (for now).

Bug: T275752
Change-Id: Id8127d19eadcabfd85edf6d6377298a64a3cf85f