maps: Disable http2 in tlsproxy
Why:
It turns out that http2 in maps was probably inadvertedly enabled. It
should be a conscious decision to enable http2 in internal services as
enabling it comes with a number of caveats one should be aware about
before enabling it

What:
Just switch to false. Since this is a service to be replaced by tegola,
make the switch simple

Bug: T275752
Change-Id: I19e33095af8bb23a9b9707d162d55c7be301e380