cloudelastic: Disable http2 in tlsproxy
Why:
It turns out that http2 was probably inadvertedly enabled by reusing the
profile. It should be a conscious decision to enable http2 in internal
services as enabling it comes with a number of caveats one should be
aware about before enabling it. Interestingly it seems http is not
exposed in this role, but regardless, even if it is, having http2 should
still happen after a weighting of the pros and cons

What:
Just switch to false.

Bug: T275752
Change-Id: Ic004377616e6f9e50ba4b92f08436cdb7aa08c06