cumin: add parsoid-canary to mw-canary and reuse other aliases
As the ticket requests parsoid hosts should be included in mw aliases
nowadays.

First doing the low-hanging fruit, add parsoid-canary to mw-canary and
use the existing aliases so mw-canary is app-canary + api-canary + parsoid-canary.

As long as just "mw" means "appservers only" and not "all of mw" this does not belong
there either though.

So I am making another suggestion to turn "mw" into "all of mw"
in a separate change.

Bug: T294802
Change-Id: Ibb37f98be81058fb3db4c0c9876a24340c9c4f89