mediawiki::php: support multiple php version in monitoring too
This means, for now, that we're exporting all metrics and checking
that the daemons are running, and the status of opcache. We still only
monitor a single php engine at the moment as we lack the web
configuration to reach any php version besides the default one.

We do separate, however, prometheus metrics by version.

Bug: T293450
Change-Id: I2358d292706c1ba86f81abf83823b7c6991b553a