Add sslcert::trusted_root_ca
This is a little helper to deploy the chained certificate
of the WMF's interal root CAs (PKI and Puppet).
The define by default deploys the .pem file in a defined location,
and it optionally create a .p12 truststore file containing it
(to be used for various use cases, for example JVM-based TLS
deployments).
The .jks truststore is not provided yet since most of the Java
systems that we run accept .p12 files. It is also nice to provide
a dependency to openssl only, rather than openjdk (to have
the keytool command).

Bug: T291905
Change-Id: I2639ed2015643a66c54d3c1d863f430a4fe2a682