profile::kafka::broker: allow to override super_users
In order to allow a smooth transition between puppet host TLS certs
and PKI TLS certs, we shoudl support the use case of a cluster
with some brokers running the new TLS certs and some others
with the old ones. We should allow to override the super_users
via hiera, so that before upgrading any broker we can change
the list of allowed CNs.

Ideally the $use_pki_super_users flag should be turned to true
as pre-step before switching the cluster to the new TLS certs,
so that all the brokers will allow more CNs to be super users.

This flag is temporary, until we migrate all clusters to the
new PKI intermediate CA.

Bug: T291905
Change-Id: Id10e3dfbb6ef7606d514b1e948ec2abc226ef8e7