snapshot:: add a 'silent' flag to fulldumps.sh to run it from timers
If parameter $6 is set to 'silent' then do a run that does not
output, like a dryrun, but actually does run.

This is to be used in I5fc9a7fe45612a to avoid sending mail every
time without being able to redirect the output to /dev/null as before
when we convert from crons to systemd timers.

Bug: T273673
Change-Id: I70fa9820fdd6d2dc359112f7c67f2d8a1c6e2835