profile::kafka::broker: add conditions to truststore deployments
Before starting the TLS certificate swap, all the brokers need
to update their settings to trust both the Puppet CA and the Root PKI
CA in their truststores (in addition to trust more CN: values).

This change adds a specific hiera value for the pki migration,
that takes care of all the pre-step to be done.

Set also the truststore password, since it seems that Java doesn't
like an empty value.

Bug: T291905
Change-Id: Ib1fc63e529a37dec84c79bba85d110e8560d06c4