bird::anycast_healthchecker: allow customization of logging options
On hosts that are calling profile::bird::anycast, anycast-healthchecker
is configured and that also includes setting up logging for it.

Currently, in the anycast-healthchecker.conf file, we set the verbosity
of logging to "info" and "log_backups" (number of old log files to
maintain; log is rotated by anycast-hc itself) to 8.  It's possible that
for cases where the host is already running and anycast-hc configured
correctly, we may not need the "info" log level and something like
"error" or "critical" should suffice. This will help us avoid logging
extra information/logs to disk that we may not need and will also help
us to conserve disk space.

As an example on the doh1001 host:

    62M	        /var/log/anycast-healthchecker/anycast-healthchecker.log
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.1
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.2
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.3
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.4
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.5
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.6
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.7
    101M	/var/log/anycast-healthchecker/anycast-healthchecker.log.8
    863M	total

This commit allows configuring the logging level and the number of old
logs that are retained. If desired, hosts that are calling bird::anycast
can configure the log level and also the number of old logs they wish to
retain. For something like doh*, we will probably set it to:

    profile::bird::anycasthc_logging:
        level: 'critical'
        num_backups: 2

In a later commit and if this is acceptable, we can extra customization
options to the same type, such as the size of the logs, sending them to
a central syslog server, etc.

Change-Id: I9f81e79577e31495952cbc083d78ebac1fbba9ac