scap: Limit foreachwikiindblist and expanddblist in beta to beta wikis
In 2013, a similar change was made to foreachwiki where commit
Ib0f0f1851e (a3f8c1281c) swapped the singular "all" argument for "all-labs".

This was important at the time, because the Beta Cluster had `en_rtlwiki`
and `deploymentwiki` which existed in Beta, but not in production.
This is no longer the case in 2025 (I65f465d1ce, and T198673).

In 2021, it was noticed that scripts that use foreachwikiindblist
or that pipe expanddblist directly are broken in Beta. Fix this
by fixing this at the source in expanddblist instead.

Bug: T357877
Hosts: deploy1003.eqiad.wmnet
Change-Id: I9f6a51da5ffc375488c79ac56e66ae29b4519d63