profile::base::certificates: add sslcert::trusted_ca options
In a lot of use cases daemons/clients need to trust either the Puppet CA
or the Root PKI CA when connecting to an endpoint. In some use cases,
like jvm-based deployments, a .p12 or .jks truststore is needed.

This change converts the sslcert::trusted_ca define to a class,
and deploys it via profile::base::certificates. The idea is to:
1) Always deploy the /etc/ssl/localcert/trusted_ca.pem bundle.
2) Optionally create a .p12 or a .jks truststore out of the .pem
   bundle. Since in the jks use case some openjdk dependencies
   are needed, it is better to keep it disabled by default.

Immediate use case: the Kafka clusters are transitioning from
Puppet-CA-based TLS certs to a dedicated PKI Intermediate CA,
but to guarantee a smooth transition clients and brokers need
to trust both CAs at the same time.

Bug: T291905
Change-Id: I5f58f9daa3841a4919d3ca96912216ef14f06066