varnish::logging: pass list of mtail programs from profile
The mtail programs to be installed are awkwardly listed in the varnish
module. Plus, we use two different mechanisms to specify
cluster-specific programs. varnishmedia, which is specific to
cache_upload, is installed with a $cache_cluster conditional in the
module, while varnishrls, specific to cache_text, is included by the
profile cache::varnish::frontend::text.

This commit moves the logic that decides which programs to install to
the profile, based on the cache cluster the host belongs to. The list of
programs is then passed to varnish::logging.

By doing so we follow the module/profile/role paradigm more closely and
make things a bit simpler. Also, we can now get rid of the cache_cluster
parameter from the varnish module, as well as varnish::logging::rls and
varnish::logging::media.

Bug: T293879
Hosts: cp3050.esams.wmnet, cp3051.esams.wmnet
Change-Id: I5aa6c014e1b87baf3f210e23bcab8c6aff874510