Update the times at which refine_sanitize monitor jobs are run
We experienced issue with the times at which the monitoring jobs for
refine santization (delayed) jobs were running, particularly in relation
to new schemas that were being backfilled.

This change ensures that the monitor_interval parameter to the
refine_job defined type is required to be set when defining the
resource. It also sets this parameter for each of the 14 instances of
this type, with a sensible schedule.

Change-Id: I2660824ed5e31891eb7611714f23699ab7e47abe