varnish: Mimick XFF behaviour with UDS + PROXY protocol
The applayer expects the cache node IP on the XFF header.
Ensure that varnish adds its own IPv4 address when
UDS + PROXY protocol is used. This is required because
varnish appends client.ip to XFF by default and client.ip
is set to the real user IP when PROXY protocol is used

Bug: T290005
Change-Id: Id6d8a8527ef6bcaadfd79668d637560453fd3fd8