gitlab: turn on Content-Security-Policy
We had report generate for a few weeks now and it seems the sole report
we get originate from Firefox extension being prevented from injecting
code which is a bug in Firefox.

I have spotted one case of font src being blocked, but it might be an
extension as well.

I have enough confidence to enforce the CSP now. This is done on both
the primary and the replica, there is not much need to do a rolling
deployment since we have the reports already.

Bug: T285363
Change-Id: I859c50fe992c909d9c9d92e97f05f6c2f8dc8be1