cloud/devtools: sync Horizon Hiera with repo Hiera
The goal here was that in the devtools project we do not
mix different Hiera locations like we ended up doing in deployment-prep.

Though things keep being added in the web UI, some are duplicate and
are already in here, others were not.

Syncing this and removing from web UI what is in the repo, as it was
intended here.

Bug: T294174
Change-Id: I6303adbb22f4ef72c8a0f6fb57c10700b4c44cb1