dynamicproxy: add tls to api
Note that this uses the old HTTP port for HTTPS - I don't want to
manually change the neutron security rules for each allowed host. This
requires that the URL on Keystone is changed roughly at the same time
this is merged.

The read-only port was not changed. There are too many users where we
can't control the protocol and soon these separate ports can be merged
anyways.

Change-Id: Ib00eee8bc20e957e75e37a50a4c495edcd5184a9