cassandra: move cluster:user relation from 1:1 relation to a 1:many
Currently we assume that one cluster will have one user and one
keyspace. This change is an attempt to change that relation, allowing
one cluster to have multiple users defined. We don't currently codify
grants in puppet but in templates. Rather than use the same grants for
all users, define a grants file on a per-user basis. Currently all
grants files are the same because I don't want to change any existing
permissions in this CR at least.

Additional work will be required in the labs-private and private repos
in order to populate the profile::cassandra::user_credentials
dictionary to be of the format
```
user_credentials:
  restbase: blahblah
  aqs: blahblah2
```
etc

Bug: T235299
Change-Id: I6284784d61a6587885fd8139d3bf5929f08b4444