varnish: add varnish::logging::mtail
Move the logic for installing and configuring a varnishmtail instance to
its own define. Do not add or remove any instance yet, keep on
configuring 'default' pretty much as before.

The main difference is that we now install the mtail programs under
/etc/mtail-default/ instead of /etc/mtail/ and that we explicitly pass
the mtail port number as the second argument of varnishmtail-default.
Also, passing both the mtail programs directory and port number is now a
requirement.

Bug: T293879
Hosts: cp3050.esams.wmnet, cp3051.esams.wmnet
Change-Id: Iad10e4f30ff85275422256261eab4a724abcf0ae