zuul: send errors from git-daemon to client
git-daemon defaults to not give any details back to the client when it
encounters an error and would thus come up with a generic message:
"access denied or repository not exported":

  static int daemon_error(const char *dir, const char *msg)
  {
    if (!informative_errors)
      msg = "access denied or repository not exported";
    packet_write_fmt(1, "ERR %s: %s", msg, dir);
    return -1;
  }

From git-daemon(1) the intent is to avoid leaking "information about the
existineence of unexported repositories". We have `--export-all` so we
really export everything and can turn verbose messages.

That might help debugging fetches failling apparently cause a merge
operation is being done on the repository while it is being fetched.

Bug: T187897
Hosts: contint2001.wikimedia.org
Change-Id: Ib4dce745acfa2dded47d9e2e96b0b85146a85aaf