scap::target - add $manage_ssh_key parameter
There are cases where deploy_user is managed outside of scap::target,
but the ssh key only exists for use by scap. Adding this
parameter allows manage_user to be false, but manage_ssh_key to be
true.

$manage_ssh_key defaults to the value of $manage_user, so
this should be a no-op everywhere.

Set params accordingly for airflow scap targets.

Bug: T295380
Change-Id: I3d28b8490bb9622e484fbdcc860d5461f6207c73