cloud: ceph: migrate admin keyring to new abstraction
This patch drops all code related to manage the ceph admin keyring as an
independent entity, because we're moving into a new abstraction that
introduces a more convenient way to manage ceph keyrings.

So this is a 3 step process:
 * first we drop the code to manage the keyring
 * include keydata in the puppet private repo
 * then we enable keyring management via the new abstraction

This patch deals with the first step.

To clarify usage of the admin keyring, it is being used by
 * mons / osds ceph servers
 * backy2 @ some cloudvirts
 * backy2 @ cloudcontrol for glance images

Bug: T293752
Change-Id: Ib97f6d3615075e569c05a17581602adfc2eb50a2
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Signed-off-by: David Caro <dcaro@wikimedia.org>