logstash: reconstruct gitlab sidekiq message field
The format appears to have changed and the "status" field is now
text indicating job state which conflicts with the rest of the gitlab suite.

Leave the status field transformation the same for other gitlab apps, but move
it for sidekiq only into the message field to be closer to the previous format.

Bug: T295731
Change-Id: Ifa644f0d21823892d694d2566bffdfcb32c723dc