ci: upgrade lintian-junit-report
Jenkins Debian glue transform the lintian text report to a Junit XML
report. It is however not recognized by the Jenkins Xunit plugin due to
`<testsuite>` elements missing a `name` attribute.

The fix I have made is in Jenkins Debian Glue v0.20.0 but we do not have
it on Buster. Until we upgrade the CI agents to Bullseye, monkey patch
lintian-junit-report.

Bug: T295719
Change-Id: Id4d6e5438eb3b5016da769eee2d9dc7d2d1c19e0