acme_chief: convert cron to restart service to timer
I was reminded by cron spam from the acmechief-test machine
that this is still a cron.

And we want to convert all crons to timers anyways.

Assuming you want email about failures but suggesting to
send to sre-traffic@ instead of it becoming root mail.

Does that make sense?

Still running hourly but not even caring about calendar or the minute,
just making sure it happens hourly from when the service started.

This will absent the crons and a follow-up change would then remove the
cron code.

I guess the TODO is still valid and not related to this.

Bug: T273673
Change-Id: I8f987d80dca384ce34c72fece00d482f50a8cd56