gitlab-runners: move profile::gitlab::runner::docker_volume: true to repo
Checked in Horizon and every runner has this in Hiera, but added in
the web UI.

Since it looks like this should always be on let's just add it here
in the repo, in the existing file.

Then we never have to worry about manually adding that in Horizon
when creating new runners in the future and everything is in one place.

Would then remove from Horizon after merging here.

Change-Id: I3faf5b1c6dc74f330e012be9e9260f9849d67545