profile::mediawiki::php: support kubernetes in php-fatal-error.php
On kubernetes, the php container doesn't have access to the system-wide
syslog, but can submit its logs to a udp listener active inside the pod.

In that case, manually format the message to send and send it over the
socket. Errors are suppressed as this is already the handler for wmerrors.

Bug: T288851
Change-Id: Icf3d9ae97d997b3d3bbdcb3a4f3be59948695390