admin: add mbsantos and jgiannelos to parsoid-test-admins
As requested on ticket, both Mateus and Yiannis are new to
the content transform team (parsoid) and should have access
to the parsoid test servers.

Subbu is the approver and I had direct approval from him on IRC
and he created the ticket:)

Since these are existing shell users added to an existing group
we don't additional steps and in the interest of time I'll go
ahead and merge this.

Bug: T295900
Change-Id: If6164575e7ce4a8368d6d001f43019669fe6be7f