admin: upgrade ihurbain from ldap_only to shell, add to parsoid-test-admins
Isabelle is new to the content transform team and should
have shell access to scandium and testreduce1001, parsoid test hosts.

The parsoid-test-admins group does this and I just added Mateus
and Yiannis to it.

Requestor on ticket is also approver for the group and this is
adding users to an existing group so that should be good to go.

Bug: T295900
Change-Id: I300c233b38d5882b1a9c770d0c22b05c77e79eb3