profile::cache::kafka::webrequest: add pki settings
To ease the migration of the Kafka Jumbo Brokers of their TLS certs
from the Puppet CA to the new PKI Kafka Intermediate, we need to
point varnishkafka producers to a new bundle that contains both
CA certificates (deployed by profile::base::certificates on all
hosts).
This change enables the new settings for cp3050 to test it properly
(even if on other librdkafka-based tools it worked perfectly).

Bug: T291905
Change-Id: I4ff0104d6c26640f9bd698440de5bc4a56b06de7