wikidough: set CSP headers for the landing page
In 021e0cf0, we first added CSP headers for durum as it will likely be
more user-facing than the landing page for Wikidough. Even though most
users will just use wikimedia-dns.org/dns-query in their browser for the
DoH resolver, nevertheless, we should set CSP headers for the same.

default-src is set to 'none' as a safe default. We set style-src and
script-src to self to accommodate branding. frame-src could also be
'self' but there doesn't seem to be any need for it; frame-ancestors
should be set to 'none'.

Hosts: O:wikidough
Change-Id: I63063254d7d89ed2a406ce5506a32386678a605f