cloudbackup: require the lvm package
To install the package, otherwise the volumes wont get created.

This will likely conflict with the other package declaration in the old
nfs backup role, so while at it, stop loading the cinder backup profile
into the nfs backup servers.

This role should be applied now only to the cloudbackup1001-dev server,
we no longer need to run cinder-backup on cloudbackup2002.

Bug: T295584
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I59b22d19f09f2d20b40b233a73d2f983e59efee7