superset: set webserver timeout to 180 seconds
We previously increased the `SQLLAB_TIMEOUT` to 3 minutes, but dashboard
queries still time out after 1 minute because of the
`SUPERSET_WEBSERVER_TIMEOUT` setting. This sets them both to 3 minutes.

Since gunicorn is the upstream server, it needs a timeout that is
longer; I chose 3 minutes 5 seconds.

Hosts: an-tool1010.eqiad.wmnet
Bug: T294771
Change-Id: If5c356de8fed351d2ab4099ce7b28b92303ada59