admin: Add ksiebert to analytic privatedata group
Adding the aforementioned user in to the group
analytics_privatedata_users will ensure that they are able to
access private data in the pre-defined dashboards (hive/superset).
The process is further explained here https://w.wiki/4SRN

User was also removed from the ldap_only_user list and added to
users list as initial CI test was failing. For a user to exist
in a group, they must first have an entry in the user list.

The Task T292053 further details as to why the user was initially
a member of ldap_only_users.

Bug: T295777
Change-Id: Ife1a2d5ae409a0e5c4aac2ab375bf531520e7733