mariadb: Split the dbstore_multiinstance role into two others
dbstore_multiinstance was not really a role, but a functional
profile- it meant what was needed on the host, not why.

dbstore_multiinstance was being used into 2 separete kind of
hosts that used to be the same: backup sources and analytics
replicas. Split those into:

* mariadb::backup_source: to generate backups for recovery
                          of mediawiki hosts
* mariadb::analytics_replica: used by analytics and researchers
                              to run queries on a separate
                              network

This requires, aside from creating the new profiles, a few
hiera changes and other references.

For now, the profile will be kept the same, but could change
in the future, and hiera keys could be done differently based
on the different role.

Bug: T296285
Change-Id: I1fa1e0e901527974f7098595a791fbfc142bb8f2