thanos::frontend: correct the realserver pools services
The "services" values in profile::lvs::realserver::pools should be
"all services that are needed on the local machine needed to respond
to requests directed to this lvs pool".

That will produce a restart-<service_name> script that will depool the
server by all the relevant lvs pools before performing a service restart.

Change-Id: I0a453d3e4bbcd22645fdaf36631168439cb3ea9a