admin: Add user taavi to wmcs and labtest group
User has been associated with wmcs-roots and labtest-roots group.
In addition, they have been added to the users list and removed from
the ldap_only_users list so as to be associated with the two groups
mentioned above.

The change also captures persons responsible for approval for the
two groups above.

Bug: T296192
Change-Id: I81e9d23defaef8bc5c490b02eebd13a46cc85665