P:base::certificates: update support for trusted CA
Currently in production we distribute a package of trusted CA which
includes a trusted CA bundle to allow services to more tightly tune what
is allowed to connect.  In cloud environments we have no such package as
such we need to build the trusted bundle from files already on the
system

This CR updates the code so that we have can provice either set of files
to build the trusted CA.  or a packaged along with some meta data
explaining where the trusted CA us.  this allows us the greatest
flexability in supporting multiple environments and at the same time
gives us the information needed to build java truststores

Bug: T296089
Hosts: kafka-test1006.eqiad.wmnet
Change-Id: If2ad4ecb1db938b91f21e3e4508d863593ced354