Add drmrs public prefix to ntp allowed config
The IPv4 range 185.15.58.0/24 is allocated to our new drmrs site.
NTP requests from mr1-drmrs to any of the dns servers are currently
failing, however, as this is not defined in ntp.conf as an allowed
client range.

This change should alter that and allow mr1-drmrs to sucessfully get
time from the dns hosts.

Bug: T296623
Change-Id: I0eb9a8315ae164280239448dfffedfb69db58b3f