ceph: auth: introduce new parameter 'import_to_ceph'
This new parameter tells the code logic if the keyring should be imported
into ceph. This is useful in a later patch, in which we store dummy keyrings
that shouldn't be imported into ceph.

Disclaimer: we have yet to decide if we want to store such 'dummy' keyrings in our
hiera data structure, or we want to generate them on the fly instead.

Bug: T293752
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ibe89a600d5a71f2a0bb919afdccffaea42fd029b