Add /srv/spark-tmp to the list of allowed read-write paths
We have recently added a default option to spark to create temporary
files under /srv/spark-tmp on the stat100x servers. This worked in the
console, but failed for jupyterhub users.

The reason for this is that the systemd unit created by
jupyterhub-conda-singleuser has a defined list of paths for which
read/write access is permitted.

This commit aims to correct that error by adding /srv/spark-tmp to the
list of permitted read/write paths when creating the service.

Bug: T295346
Change-Id: Ib96f14562fc96317d8a9c3a0f3fe00da17a1e0bf