Refactor superset caching to enable dual caches
Superset has two discrete caches, one for its own metadata and another
for the results of data queries. We wish to re-enable the metadata
cache, whilst leaving the data cache disabled (because it doesn't
observe necessary security restrictions).

This commit changes the single parameter into two discrete parameters
and enables the metadata cache for superset in production and staging.

Bug: T295295
Change-Id: I1fd0286155b3a5bf14da430ebf504dfd35f3a8a6