durum: show if the user is using DoH or DoT
The durum check service allows a user to confirm if they are using
Wikidough or not. So far, this has only been a simple "yes" or "no". To
make a distinction between DoH and DoT, and to clearly indicate which
service a user is using, this commit adds support for indicating the
type of service (DoH or DoT) a user has configured in Wikidough.

This adds a new IP address that is anycasted and that we listen on:
yesdot. The results of the check page are also updated.

Hosts: O:durum
Change-Id: If138c44b9a02ebe7f5c88c3210210753441cb8c9