dnsdist: refactor the configuration template for updates to durum
In preparation for the updates to the durum check service, refactor the
dnsdist.conf.erb template to add new rules for returning the desired
service IPs for the durum service. The purpose of these changes is to
allow us to distinguish if the user is using DoH or DoT and to reflect
this on the durum landing page.

Hosts: O:wikidough
Change-Id: I14533b970aba3d9d39fddbb952b20d63d55ed4bd