bird: add validate_cmd for anycast-healthchecker.conf
We should validate the configuration for the anycast-healthchecker
service and specifically anycast-healthchecker.conf. This commit adds a
validate_cmd that uses anycast-hc's --check option so that we have a
Puppet failure and not an anycast-hc service failure in case of a
missing/incorrect configuration.

Note that this check does not cover all cases (such as duplicate
vip_fqdn's or old ones), but it will still be nice to have it.

Hosts: O:durum
Change-Id: Ib213130eeeaf5fab14d3e031faf34b4c0040240a