contint: delete the proxy_gerrit class
This class and the code say this is only for cloud.

When searching in https://openstack-browser.toolforge.org/puppetclass/
I don't see it applied on an instance.

And if it was it should be included in a role/profile which is then
applied.

Does anyone know or remember where this was used? Can we delete it?

Bug: T272559
Change-Id: I6b0379c34d5b8c42146737696e39eed9731feb93