mgmt: delete the entire module and role::mgmt::drac_ilo
This was once made after the mgmt password leaked and we had
to reset it on all hosts at once.

It would take the mgmt names from DNS repo, guess the vendor from the
SSH banner, then run the right password reset command for either
Dell or HP.

But nowadays we have sre.hosts.ipmi-password-reset and that should
replace this. Does it make it sufficiently easy to change it on ALL if
ever needed again?

Bug: T272559
Change-Id: Iabc4692310ee52bffe0d0b1826ea23671279a2ac