rdf query service: limit namespace aliasing to /bigdata/namespace
The _check_auth subrequest was hitting this conditional, it seems that
the $namespace variable is inherited from the parent request and is
allowing the rewrite.

Fix by making the rewrite more narrow. The namespace should only be
set when the url is within /bigdata/namespace, so limited rewrites
to match.

Change-Id: I1048692469b7266bd05bfea4d4282b9ede3e6c85