durum: show a message about the limitations of the DoH protocol
If a user is using DoH, show a message to remind them that if DoH is
being used from a browser (as opposed to from the OS layer), that DNS
lookups are only secure from within that browser and not from other
applications.

For users who are not familiar with the protocol, giving them a false
sense of security should be avoided. In addition to this message for
DoH, we should probably indicate for both DoH and DoT the limitations of
the protocols themselves (SNI and IP is visible) but that's for later.

Hosts: O:durum
Change-Id: Iafbc2cf92e764742564b0808c8154d4619ceb766