Increase the timeout for Druid on the analytics cluster
We wish to allow for longer-running Druid queries on the analytics
cluster. The current value of 10s is proving too short for users of
Superset so we wish to raise this timeout value. We are doing so
incrementally, with the first such change being to 30 seconds.

Note that this does not affect the public Druid cluster and the
public-facing AQS services.

Bug: T297148
Change-Id: I3e981eb2881c73aa871b4b4425b6fd133cd573d5