ceph::auth::keyring: ensure parent directory tree exists
If the keyring has a weird keyring_path, the parent directory tree may need to
be created and managed by puppet.

We cannot do this in ceph::mgr because that class only recieves its own host keyring
as parameter, so the parent dir tree is created only for the host keyring, and the
others will fail.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Id7b9bed6fd197ffaae661a5089e66658b0b73a0a