annualreport: switch git::clone latest to present
In the past this would be a setup where content admins
would want to merge changes in the dpeloyment repo and get an
auto-deploy via puppet.

But nowadays this is just a redirect to wikimediafoundation.org
and new annual reports are released there.

So for the event where once a year a redirect has to be updated
it's fine if an SRE has to be pinged to git pull.

Bug: T218900
Change-Id: Ie1e9feaa97cfd5f452bd318a32aa040f4ce8c55a