bienvenida.wikimedia.org: switch git::clone from latest to present
The last update in the content repo this pulls from was in 2018.
This is more like a static site since then.

So it seems safe to disable the 'latest'. It seems very unlikely
someone wants a change and if they wanted to they would notice they
can't update production and eventually contact SRE.

Bug: T218900
Change-Id: I5fac3772a6baff61a291684454aec8a7e7d35b13