transparency.wikimedia.org: switch git::clone from latest to present
After T230638 this is another microsite that once had content deployers
but was replaced with Wordpress on an external server it merely
redirects to.

The last edit in this content repo was in 2020 related to archiving
the site as historic.

https://gerrit.wikimedia.org/r/q/project:wikimedia/TransparencyReport

Therefore we also don't have to expect anymore that the comms
team might want to change this. At least not without having to
ask us (anymore) and we set this to "present" as well.

Bug: T218900
Change-Id: I039f25a832f26370fb078dd4bdd996ad3dc79131