wmflib: add service::get_services_for function
We want to centralize the logic around common fetch actions on the
service catalog, so we don't need to repeat ourselves in different parts
of the repository.

Also: change the logic for discovery records, so that they appear in the
monitoring_setup phase as many checks might depend on the discovery
record being present.

Change-Id: I780634d5207e84e1716b25f8dd10ca26437df17d