prometheus: pin discovery probes to their site
Add logic to find the service's name in the local 'svc' zone. This name
is used to reach the service over the network, and send the
discovery.wmnet hostname as SNI/Host header.

The logic might not cover all cases, for historical/legacy services we
might need to adjust the 'aliases' array to include the correct name.

Bug: T291946
Change-Id: I6908e1de1cf3a63b135780d435c44776da127aec