mediabackups: Make wiki optional, add another optional parameter: dblist
Add a new parameter to multimedia backups: dblist, this will point
to a file name inside the repo:
operations/mediawiki-config:mediawiki-config/dblists

eg. s1.dblist, all.dblist

Make the wiki parameter optional, alongside with the new- normally
you would define one or the other (a wiki, or a list of wikis),
although it is possible to define both (wiki will be ignored) or
none (an error will happen).

Bug: T262668
Change-Id: Ibc9b0e01eb12c53dda165e005bb00566be741d53