mediabackup: Add an encryption key to store private files securely
An encryption key will be setup on the private repo (a different
one for each datacenter) in order to encrypt and decrypt non-public
files (files from closed, private or deleted wikis) in order to
prevent leaks.

The code uses the encryption key with 'age' package to encrypt
and decrypt the files. The file available on the hosts will also
be added to the configuration.

Bug: T262668
Change-Id: I5e33ba8c1378bfa5998d31a26dd683c771012308