puppetmaster: Install 'age' on puppetmaster frontends
age is a modern alternative to gpg which can be used to easily
encrypt and decrypt files, as well as generate keys for those.
It can also use ssh keys for encryption.

Install it on puppetmasters so it can handle media backups keys
in the private repo, rather than generate them elsewhere and
copy them.

Bug: T262668
Change-Id: Id9191c389e96441667713baaaf8ea23373ae32d4