gitlab_runner: use config template for registering new runners
This change introduces --template-config flag for registering
new runners. This config templates allow to configure at least
[[runners]] config section.
Global configuration can not be configured with this templates
and is implemented with file_line currently. GitLab Runner
alters the config.toml file when registering new runners, so
this is a compromise to make it somewhat puppet-managed.

Bug: T295481
Change-Id: I422499327957f0efad9f2d1de7737de5e59100fb