sonofgridengine: grid-configurator: introduce more code to cleanup dead config
Dead configuration is produced when a VM instance is deleted from openstack but
some files referencing it still exists on the grid configuration.

The grid-configurator already has some logic to handle such cases, but I've
confirmed that at least the 2 cases this patch covers were not previously handled
well, and prevented grid-configurator from properly operating, because DNS
resolutions will fail.

Such cases are:
 * host configuration files in directory /var/lib/gridengine/etc/hosts/random_vm.eqiad1.wmcloud
 * hostlist parameter in hostgroups config file
 * hostlist parameter in queues config file

Example of scenario that makes this fail:

 $ sudo /usr/local/bin/grid-configurator --domain queue --beta --debug
  DEBUG Running configuration updates for queue
  DEBUG /data/project/.system_sge/gridengine/etc/queues/task matches task
  DEBUG /data/project/.system_sge/gridengine/etc/queues/continuous matches continuous
  root@toolsbeta-sgegrid-master.toolsbeta.eqiad1.wikimedia.cloud modified "webgrid-lighttpd" [..]
  unable to resolve host "toolsbeta-sgewebgrid-generic-09-3.toolsbeta.eqiad1.wikimedia.cloud"

The 'unable to resolve host' error prevents the reconfiguration operation from succeeded.
In this case, the mentioned hosts had dead config, and this patch knows now how to handle it.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I428843bd90c00f8ebecd0e0cf71eb359a57fff3f