add static-tendril.wikimedia.org ServerName to its apache site
We call this apache site "static-tendril" but so far the actual
ServerNames/Alias we have added are tendril and dbtree, not static-tendril.

But I expect sooner or later all 3 are supposed to work and point to
the same archive page we are setting up here.

Also this should fix the current httpbb test.

Make static-tendril the main server name and add tendril and dbtree
as aliases.

In I30cd749b66c92 this was also added to the TLS cert.

Bug: T297605
Change-Id: Ib7e7353853e8558907f3829483cef1e030798389